mubleriavasquez_backend/backend/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==3.0.9
uWSGI==2.018
django-cors-headers==3.4.0
django-import-export==2.3.0
djangorestframework==3.11.1
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
psycopg2==2.8.5
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
virtualenv==20.0.31
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
